Move gifting and corporate code into separate packages
