+++ support-frontend/app/wiring/Services.scala
@@ -8,7 +8,7 @@ import com.gu.support.config.TouchPointEnvironments
 import com.gu.support.getaddressio.GetAddressIOService
 import com.gu.support.pricing.PriceSummaryServiceProvider
 import com.gu.support.promotions.PromotionServiceProvider
-import com.gu.support.redemption.RedemptionTable
+import com.gu.support.redemption.corporate.RedemptionTable
 import controllers.DynamoTableAsyncForUser
 import play.api.BuiltInComponentsFromContext
 import play.api.libs.ws.ahc.AhcWSComponents
