+++ support-workers/src/main/scala/com/gu/zuora/subscriptionBuilders/DigitalSubscriptionBuilder.scala
@@ -9,10 +9,10 @@ import com.gu.monitoring.SafeLogger
 import com.gu.support.catalog.ProductRatePlanId
 import com.gu.support.config.{TouchPointEnvironment, ZuoraDigitalPackConfig}
 import com.gu.support.promotions.{PromoCode, PromoError, PromotionService}
-import com.gu.support.redemption.GetCodeStatus
-import com.gu.support.redemption.GetCodeStatus.{InvalidReaderType, RedemptionInvalid}
-import com.gu.support.redemption.generator.CodeBuilder.GiftCode
-import com.gu.support.redemption.generator.GiftCodeGeneratorService
+import com.gu.support.redemption.corporate.GetCodeStatus.{InvalidReaderType, RedemptionInvalid}
+import com.gu.support.redemption.corporate.GetCodeStatus
+import com.gu.support.redemption.gifting.generator.CodeBuilder.GiftCode
+import com.gu.support.redemption.gifting.generator.GiftCodeGeneratorService
 import com.gu.support.redemptions.{RedemptionCode, RedemptionData}
 import com.gu.support.workers.ProductTypeRatePlans._
 import com.gu.support.workers.lambdas.DigitalSubscriptionGiftRedemption
