+++ support-workers/src/test/scala/com/gu/support/workers/JsonFixtures.scala
@@ -7,7 +7,7 @@ import com.gu.i18n.Currency
 import com.gu.i18n.Currency.GBP
 import com.gu.salesforce.Fixtures.{emailAddress, idId}
 import com.gu.support.promotions.PromoCode
-import com.gu.support.redemption.generator.CodeBuilder.GiftCode
+import com.gu.support.redemption.gifting.generator.CodeBuilder.GiftCode
 import com.gu.support.workers.encoding.Conversions.StringInputStreamConversions
 import io.circe.syntax._
 import org.joda.time.{DateTimeZone, LocalDate}
