+++ support-workers/src/test/scala/com/gu/zuora/subscriptionBuilders/DigitalSubscriptionBuilderSpec.scala
@@ -7,9 +7,9 @@ import com.gu.i18n.Currency.GBP
 import com.gu.support.config.TouchPointEnvironments.SANDBOX
 import com.gu.support.config.ZuoraDigitalPackConfig
 import com.gu.support.promotions.{PromoError, PromotionService}
-import com.gu.support.redemption.GetCodeStatus.InvalidReaderType
-import com.gu.support.redemption.generator.GiftCodeGeneratorService
-import com.gu.support.redemption.{DynamoLookup, GetCodeStatus}
+import com.gu.support.redemption.corporate.GetCodeStatus.InvalidReaderType
+import com.gu.support.redemption.gifting.generator.GiftCodeGeneratorService
+import com.gu.support.redemption.corporate.{DynamoLookup, GetCodeStatus}
 import com.gu.support.redemptions.{RedemptionCode, RedemptionData}
 import com.gu.support.workers.lambdas.DigitalSubscriptionGiftRedemption
 import com.gu.support.workers.{DigitalPack, Monthly, Quarterly}
